wrong field size calculation in ODBC code #1659 and other develop backports

+++ Data/ODBC/src/ODBCStatementImpl.cpp
@@ -451,7 +451,7 @@ int ODBCStatementImpl::affectedRowCount() const
 			_affectedRowCount = static_cast<std::size_t>(rows);
 	}
 
-	return _affectedRowCount;
+	return static_cast<int>(_affectedRowCount);
 }
